requirements.txt
amazon-transcribe==0.6.2
anyio==4.9.0
asttokens==3.0.0
aws-sdk-signers==0.0.3
aws_sdk_bedrock_runtime==0.0.2
awscli==1.38.38
awscrt==0.16.26
boto3==1.37.38
botocore==1.37.38
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
decorator==5.2.1
docutils==0.19
dotenv==0.9.9
Events==0.5
exceptiongroup==1.2.2
executing==2.2.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ijson==3.3.0
ipython==8.18.1
ipywidgets==8.1.6
jedi==0.19.2
jmespath==1.0.1
jupyterlab_widgets==3.0.14
matplotlib-inline==0.1.7
numpy==2.0.2
opensearch-py==2.8.0
pandas==2.2.3
parso==0.8.4
pickleshare==0.7.5
pillow==11.2.1
prompt_toolkit==3.0.51
pure_eval==0.2.3
pyasn1==0.6.1
PyAudio==0.2.14
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rsa==4.7.2
Rx==3.2.0
s3fs==0.4.2
s3transfer==0.11.5
setuptools==75.8.0
six==1.17.0
smithy-aws-core==0.0.1
smithy-aws-event-stream==0.0.1
smithy-core==0.0.2
smithy-http==0.0.1
smithy-json==0.0.1
sniffio==1.3.1
stack-data==0.6.3
tavily-python==0.7.0
tiktoken==0.9.0
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
wheel==0.45.1
widgetsnbextension==4.0.14
